Integrate js python

+++ mern/server/controllers/aiModelController.js
@@ -1,9 +1,14 @@
 import Model from '../models/aiModel.js';
 import User from '../models/userModel.js';
+import logger from '../logger.js';
+import mongoose from 'mongoose';
 
 export const createModel = async (req, res) => {
-    const { model_address, metaData_address, model_description, model_type, user } = req.body;
-    
+
+    logger.logRequestDetails(req);
+
+    const { model_name, model_address, metaData_address, model_description, model_type, user } = req.body;
+
     try {
         // Optionally check if the referenced user exists
         const existingUser = await User.findById(user);
@@ -12,6 +17,7 @@ export const createModel = async (req, res) => {
         }
 
         const model = new Model({
+            model_name,
             model_address,
             metaData_address,
             model_description,
@@ -27,6 +33,8 @@ export const createModel = async (req, res) => {
 
 // Retrieve all models for a specific user
 export const getModelsByUser = async (req, res) => {
+    logger.logRequestDetails(req);
+
     const userId = req.params.userId;
     try {
         const models = await Model.find({ user: userId });
@@ -38,14 +46,19 @@ export const getModelsByUser = async (req, res) => {
 
 // Update a model
 export const updateModel = async (req, res) => {
+    logger.logRequestDetails(req);
+
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ error: 'No such model' })
+    }
     const { model_address, metaData_address, model_description, model_type } = req.body;
     try {
-        const model = await Model.findByIdAndUpdate(id, { 
-            model_address, 
-            metaData_address, 
-            model_description, 
-            model_type 
+        const model = await Model.findByIdAndUpdate(id, {
+            model_address,
+            metaData_address,
+            model_description,
+            model_type
         }, { new: true });
         if (!model) {
             return res.status(404).json({ error: "Model not found" });
@@ -58,7 +71,12 @@ export const updateModel = async (req, res) => {
 
 // Delete a model
 export const deleteModel = async (req, res) => {
+    logger.logRequestDetails(req);
+
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ error: 'No such model' })
+    }
     try {
         const model = await Model.findByIdAndDelete(id);
         if (!model) {
@@ -72,7 +90,12 @@ export const deleteModel = async (req, res) => {
 
 // Get a single model by its ID
 export const getModelById = async (req, res) => {
+    logger.logRequestDetails(req);
+
     const { id } = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({error: 'No such model'})
+    }
     try {
         const model = await Model.findById(id);
         if (!model) {
